Refactored API items to ContentHolder from App
